caching of rotated images to help frames
